Optimise power of floating point calls
Calculating something to the power of a float is slow (i.e. `powf` is
slow). Re-arrange the math to do a power of for a larger integer and
then take the nth-root. Results come out slightly differently due to
floating-point rounding, but is sufficently close (and possibly more
mathematically correct?).

Basically:

    a^(m+n) = a^m * a^n
    (a^m)^n = a^mn

Therefore:

    a^7 = a^(3.5*2) = (a^3.5)^2
    sqrt(a^7) = a^3.5